Centralize cycle finding algorithm
