+++ tsplp/CMakeLists.txt
@@ -15,55 +15,54 @@ target_link_libraries(tsplp
 
 target_precompile_headers(tsplp
 	PUBLIC
-		<chrono>
-		<cstddef>
-		<exception>
-		<functional>
-		<limits>
-		<memory>
-		<span>
-		<tuple>
-		<unordered_map>
-		<vector>
-		<xtensor/xtensor.hpp>
+	<chrono>
+	<cstddef>
+	<exception>
+	<functional>
+	<limits>
+	<memory>
+	<span>
+	<tuple>
+	<unordered_map>
+	<vector>
+	<xtensor/xtensor.hpp>
 	PRIVATE
-		<ClpSimplex.hpp>
-		<CoinFinite.hpp>
+	<ClpSimplex.hpp>
+	<CoinFinite.hpp>
 
-		<boost/core/bit.hpp>
-		<boost/graph/adjacency_list.hpp>
-		<boost/graph/adjacency_matrix.hpp>
-		<boost/graph/boykov_kolmogorov_max_flow.hpp>
-		<boost/graph/connected_components.hpp>
-		<boost/graph/depth_first_search.hpp>
-		<boost/graph/filtered_graph.hpp>
-		<boost/graph/one_bit_color_map.hpp>
-		<boost/graph/stoer_wagner_min_cut.hpp>
-		<boost/graph/topological_sort.hpp>
-		<boost/graph/transitive_closure.hpp>
-		<boost/property_map/function_property_map.hpp>
-		<boost/range/adaptor/reversed.hpp>
-		<boost/range/iterator_range.hpp>
+	<boost/core/bit.hpp>
+	<boost/graph/adjacency_list.hpp>
+	<boost/graph/adjacency_matrix.hpp>
+	<boost/graph/boykov_kolmogorov_max_flow.hpp>
+	<boost/graph/connected_components.hpp>
+	<boost/graph/filtered_graph.hpp>
+	<boost/graph/one_bit_color_map.hpp>
+	<boost/graph/stoer_wagner_min_cut.hpp>
+	<boost/graph/topological_sort.hpp>
+	<boost/graph/transitive_closure.hpp>
+	<boost/property_map/function_property_map.hpp>
+	<boost/range/adaptor/reversed.hpp>
+	<boost/range/iterator_range.hpp>
 
-		<xtensor/xadapt.hpp>
-		<xtensor/xarray.hpp>
-		<xtensor/xindex_view.hpp>
-		<xtensor/xmanipulation.hpp>
-		<xtensor/xvectorize.hpp>
-		<xtensor/xview.hpp>
+	<xtensor/xadapt.hpp>
+	<xtensor/xarray.hpp>
+	<xtensor/xindex_view.hpp>
+	<xtensor/xmanipulation.hpp>
+	<xtensor/xvectorize.hpp>
+	<xtensor/xview.hpp>
 
-		<algorithm>
-		<cassert>
-		<cmath>
-		<condition_variable>
-		<future>
-		<limits>
-		<mutex>
-		<optional>
-		<span>
-		<stdexcept>
-		<thread>
-		<unordered_set>
+	<algorithm>
+	<cassert>
+	<cmath>
+	<condition_variable>
+	<future>
+	<limits>
+	<mutex>
+	<optional>
+	<span>
+	<stdexcept>
+	<thread>
+	<unordered_set>
 )
 
 add_subdirectory(test)
